// src/component/interfaces/FormData.ts
export interface PersonalInfoData {
  name: string;
  surname: string;
  birthDay: string;
  birthMonth: string;
  birthYear: string;
  nationalityCountry: string;
  documentType: string;
  documentNumber: string;
  residenceCountry: string;
  telephonePrefix: string;
  telephoneNumber: string;
  address: string;
  postalCode: string;
  isPublicOfficer: string;
  currentPosition: string;
  currentEntity: string;
  positionType: string;
}
export interface LoginFormData {
  email: string;
  password: string;
}